src: rename src.c to src_common.c
Renaming src.c, as this filename will have a different purpose. File
src_common.c will contain common functions for all src types.
