+++ src/audio/src/src_common.c
@@ -38,7 +38,7 @@
 #include <stdint.h>
 #include <limits.h>
 
-#include "src.h"
+#include "src_common.h"
 #include "src_config.h"
 
 #if SRC_SHORT || CONFIG_COMP_SRC_TINY
