+++ zephyr/CMakeLists.txt
@@ -678,7 +678,7 @@ zephyr_library_sources_ifdef(CONFIG_COMP_SRC
 	${SOF_AUDIO_PATH}/src/src_generic.c
 	${SOF_AUDIO_PATH}/src/src_hifi3.c
 	${SOF_AUDIO_PATH}/src/src_hifi4.c
-	${SOF_AUDIO_PATH}/src/src.c
+	${SOF_AUDIO_PATH}/src/src_common.c
 	${SOF_AUDIO_PATH}/src/src_${ipc_suffix}.c
 )
